feat: several adjustments
