+++ src/common/locales/en.json
@@ -92,9 +92,9 @@
       "dataCollection": "Data Collection",
       "dataCollectionDescription": "Control whether to use providers that may store data.",
       "only": "Only",
-      "onlyDescription": "List of provider slugs to allow for this request.",
+      "onlyDescription": "List of provider slugs to allow.",
       "ignore": "Ignore",
-      "ignoreDescription": "List of provider slugs to skip for this request.",
+      "ignoreDescription": "List of provider slugs to skip.",
       "quantizations": "Quantizations",
       "quantizationsDescription": "List of quantization levels to filter by (e.g. [\"int4\", \"int8\"]).",
       "sort": "Sort",
