+++ src/main/store/store.ts
@@ -15,6 +15,7 @@ import {
 import { migrateSettingsV5toV6 } from './migrations/v5-to-v6';
 import { migrateV6ToV7 } from './migrations/v6-to-v7';
 import { migrateV7ToV8 } from './migrations/v7-to-v8';
+import { migrateV8ToV9 } from './migrations/v8-to-v9';
 import { migrateSettingsV0toV1 } from './migrations/v0-to-v1';
 import { migrateSettingsV1toV2 } from './migrations/v1-to-v2';
 import { migrateSettingsV2toV3 } from './migrations/v2-to-v3';
@@ -85,7 +86,7 @@ interface StoreSchema {
   userId?: string;
 }
 
-const CURRENT_SETTINGS_VERSION = 8;
+const CURRENT_SETTINGS_VERSION = 9;
 
 interface CustomStore<T> {
   get<K extends keyof T>(key: K): T[K] | undefined;
@@ -274,6 +275,11 @@ export class Store {
       if (settingsVersion === 7) {
         settings = migrateV7ToV8(settings);
         settingsVersion = 8;
+      }
+
+      if (settingsVersion === 8) {
+        settings = migrateV8ToV9(settings);
+        settingsVersion = 9;
       }
 
       this.store.set('settings', settings as SettingsData);
